Refactor RunwaySuface and Regulation to Enumeration

+++ src/main/java/ar/edu/utn/frba/proyecto/sigo/dto/airport/AirportDTO.java
@@ -12,5 +12,5 @@ public class AirportDTO {
     private String codeFIR;
     private String codeIATA;
     private Long regionId;
-    private Long regulationId;
+    private Integer regulationId;
 }
